Ajout d'un .env - clean code disccordManager - ajout de dépendence dotenv - ajout d'un backup
