+++ .gitignore
@@ -117,3 +117,4 @@ run/
 .vs
 .vscode
 .history
+.env
