+++ Whitelist-Je/src/main/java/ayoub/whitelistje/.envExample
@@ -1,0 +1,1 @@
+disccordBotToken=YOURFUCKINGLONGBNOTTOKKENGOESHERE
