+++ Whitelist-Je/src/main/java/ayoub/whitelistje/functions/WhitelistManager.java
